organized navigation into different folders
