+++ routes/myStack.js
@@ -1,0 +1,22 @@
+import * as React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+import HomeScreen from '../screens/homeScreen';
+import NotificationsScreen from '../screens/notificationsScreen';
+import ProfileScreen from '../screens/profileScreen';
+import SettingsScreen from '../screens/settingsScreen';
+
+const Stack = createStackNavigator();
+
+export default function MyStack() {
+  return (
+    <NavigationContainer>
+        <Stack.Navigator>
+            <Stack.Screen name="Home" component={HomeScreen}/>
+            <Stack.Screen name="Notifications" component={NotificationsScreen}/>
+            <Stack.Screen name="Profile" component={ProfileScreen} />
+            <Stack.Screen name="Settings" component={SettingsScreen} />
+        </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
