+++ screens/homeScreen.js
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { StyleSheet, Button, View } from 'react-native';
+
+export default function HomeScreen({ navigation }) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Button
+          title="Go to Profile"
+          onPress={() => navigation.navigate('Profile')}
+        />
+      </View>
+    );
+  }
