+++ screens/notificationsScreen.js
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { Button, View } from 'react-native';
+
+export default function NotificationsScreen({ navigation }) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Button
+          title="Go to Settings"
+          onPress={() => navigation.navigate('Settings')}
+        />
+        <Button title="Go back" onPress={() => navigation.goBack()} />
+      </View>
+    );
+  }
