+++ screens/profileScreen.js
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { Button, View } from 'react-native';
+
+export default function ProfileScreen({ navigation }) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Button
+          title="Go to Notifications"
+          onPress={() => navigation.navigate('Notifications')}
+        />
+        <Button title="Go back" onPress={() => navigation.goBack()} />
+      </View>
+    );
+  }
