+++ screens/settingsScreen.js
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { Button, View } from 'react-native';
+
+export default function SettingsScreen({ navigation }) {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <Button title="Go back" onPress={() => navigation.goBack()} />
+      </View>
+    );
+  }
