+++ styles/global.js
@@ -1,0 +1,18 @@
+
+import { StyleSheet } from 'react-native';
+
+export const globalStyles = StyleSheet.create({
+  titleText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#333',
+  },
+  paragraph: {
+    marginVertical: 8,
+    lineHeight: 20,
+  },
+  container: {
+    flex: 1,
+    padding: 20,
+  },
+});
